+++ test/HeadersTest.js
@@ -1,18 +1,18 @@
 'use strict';
 
-const LokiJS = require('./../src/Requests.js');
+const BramblJS = require('./../src/Requests.js');
 
-var LokiObj = new LokiJS();
+var BramblObj = new BramblJS();
 
-console.log('\n' + "----------LokiJS header tests------------" + '\n');
+console.log('\n' + "----------BramblJS header tests------------" + '\n');
 console.log("");
 console.log("Setting api key to: 'test_key'");
 console.log("Blake2b256 hash of api key (which should be set in the settings file of your bifrost node):")
-console.log(LokiObj.blakeHash('test_key'));
+console.log(BramblObj.blakeHash('test_key'));
 //Setting api key for requests
-LokiObj.setApiKey('test_key');
+BramblObj.setApiKey('test_key');
 
-LokiObj.createAssets('6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ', '6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ', 10, 'testAssets', 0, '')
+BramblObj.createAssets('6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ', '6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ', 10, 'testAssets', 0, '')
 .then(function(response) {
   console.log('----------------------------------------------------------------');
   console.log("create 10 testAssets result:");