+++ test/it/asset.js
@@ -1,13 +1,13 @@
 const assert = require("assert");
-const LokiJS = require("./../../src/Requests");
+const BramblJS = require("./../../src/Requests");
 
 describe("Asset", () => {
   before(() => {
-    lokijs = new LokiJS();
+    brambljs = new BramblJS();
   });
 
   it("should create assets", done => {
-    lokijs
+    brambljs
       .createAssets(
         "6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ",
         "6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ",
@@ -27,7 +27,7 @@ describe("Asset", () => {
   });
 
   /*it('should transfer assets', (done) => {
-        lokijs.transferAssets('6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ', 'A9vRt6hw7w4c7b4qEkQHYptpqBGpKM5MGoXyrkGCbrfb', 1, 'testAssets', 0, '')
+        brambljs.transferAssets('6sYyiTguyQ455w2dGEaNbrwkAWAEYV1Zk6FtZMknWDKQ', 'A9vRt6hw7w4c7b4qEkQHYptpqBGpKM5MGoXyrkGCbrfb', 1, 'testAssets', 0, '')
             .then((response) => {
                 console.log(response)
                 assert.equal(typeof response.result, 'object')